+++ gmusicapi/protocol/musicmanager.py
@@ -6,7 +6,6 @@ from six import raise_from
 from builtins import *  # noqa
 
 import base64
-from collections import namedtuple
 import hashlib
 import itertools
 import os
@@ -17,7 +16,6 @@ import dateutil.parser
 from decorator import decorator
 from google.protobuf.message import DecodeError
 import mutagen
-from oauth2client.client import OAuth2Credentials
 
 import json
 from gmusicapi.exceptions import CallFailure
@@ -29,38 +27,6 @@ log = utils.DynamicClientLogger(__name__)
 
 
 _android_url = 'https://android.clients.google.com/upsj/'
-
-OAuthInfo = namedtuple('OAuthInfo', 'client_id client_secret scope redirect')
-oauth = OAuthInfo(
-    '652850857958.apps.googleusercontent.com',
-    'ji1rklciNp2bfsFJnEH_i6al',
-    'https://www.googleapis.com/auth/musicmanager',
-    'urn:ietf:wg:oauth:2.0:oob'
-)
-
-
-def credentials_from_refresh_token(token):
-    # why doesn't Google provide this!?
-
-    cred_json = {"_module": "oauth2client.client",
-                 "token_expiry": "2000-01-01T00:13:37Z",  # to refresh now
-                 "access_token": 'bogus',
-                 "token_uri": "https://accounts.google.com/o/oauth2/token",
-                 "invalid": False,
-                 "token_response": {
-                     "access_token": 'bogus',
-                     "token_type": "Bearer",
-                     "expires_in": 3600,
-                     "refresh_token": token},
-                 "client_id": oauth.client_id,
-                 "id_token": None,
-                 "client_secret": oauth.client_secret,
-                 "revoke_uri": "https://accounts.google.com/o/oauth2/revoke",
-                 "_class": "OAuth2Credentials",
-                 "refresh_token": token,
-                 "user_agent": None}
-
-    return OAuth2Credentials.new_from_json(json.dumps(cred_json))
 
 
 @decorator
@@ -215,6 +181,16 @@ class UploadMetadata(MmCall):
 
         track.client_id = cls.get_track_clientid(filepath)
 
+        audio = mutagen.File(filepath, easy=True)
+
+        if audio is None:
+            raise ValueError("could not open to read metadata")
+        elif isinstance(audio, mutagen.asf.ASF):
+            # WMA entries store more info than just the value.
+            # Monkeypatch in a dict {key: value} to keep interface the same for all filetypes.
+            asf_dict = dict((k, [ve.value for ve in v]) for (k, v) in audio.tags.as_dict().items())
+            audio.tags = asf_dict
+
         extension = os.path.splitext(filepath)[1].upper()
 
         if isinstance(extension, bytes):
@@ -224,12 +200,18 @@ class UploadMetadata(MmCall):
             # Trim leading period if it exists (ie extension not empty).
             extension = extension[1:]
 
+        if isinstance(audio, mutagen.mp4.MP4) and (
+                audio.info.codec == 'alac' or audio.info.codec_description == 'ALAC'):
+            extension = 'ALAC'
+        elif isinstance(audio, mutagen.mp4.MP4) and audio.info.codec_description.startswith('AAC'):
+            extension = 'AAC'
+
         if extension.upper() == 'M4B':
             # M4B are supported by the music manager, and transcoded like normal.
             extension = 'M4A'
 
         if not hasattr(locker_pb2.Track, extension):
-            raise ValueError("unsupported filetype")
+            raise ValueError("unsupported filetype: {0} for file {1}".format(extension, filepath))
 
         track.original_content_type = getattr(locker_pb2.Track, extension)
 
@@ -241,16 +223,6 @@ class UploadMetadata(MmCall):
         track.client_date_added = 0
         track.recent_timestamp = 0
         track.rating = locker_pb2.Track.NOT_RATED  # star rating
-
-        # Populate information about the encoding.
-        audio = mutagen.File(filepath, easy=True)
-        if audio is None:
-            raise ValueError("could not open to read metadata")
-        elif isinstance(audio, mutagen.asf.ASF):
-            # WMA entries store more info than just the value.
-            # Monkeypatch in a dict {key: value} to keep interface the same for all filetypes.
-            asf_dict = dict((k, [ve.value for ve in v]) for (k, v) in audio.tags.as_dict().items())
-            audio.tags = asf_dict
 
         track.duration_millis = int(audio.info.length * 1000)
 
@@ -457,10 +429,14 @@ class GetUploadSession(MmCall):
             return (True, None)
 
         if 'errorMessage' in res:
-            # This terribly nested structure is Google's doing.
-            error_code = (res['errorMessage']['additionalInfo']
-                          ['uploader_service.GoogleRupioAdditionalInfo']
-                          ['completionInfo']['customerSpecificInfo']['ResponseCode'])
+            try:
+                # This terribly nested structure is Google's doing.
+                error_code = (res['errorMessage']['additionalInfo']
+                              ['uploader_service.GoogleRupioAdditionalInfo']['completionInfo']
+                              ['customerSpecificInfo']['ResponseCode'])
+            except KeyError:
+                # The returned nested structure is not as expected: cannot get Response Code
+                error_code = None
 
             got_session = False